backend/requirements.txt
fastapi==0.119.1
uvicorn[standard]==0.25.0
websockets==12.0
python-multipart==0.0.20
insightface==0.7.3
onnxruntime==1.23.1
opencv-python==4.8.1.78
numpy==1.24.3
Pillow==10.0.1
elasticsearch==8.10.1
python-dotenv==1.0.0
requests==2.31.0